Minior Adaptation to kmeans Local
Split the allocation of vectors to clusters into two parts. Firstly the calculation of distance, then the determining of the nearest center. This way I can run two instances of the classes in parallel, one for each image and change the properties externally so they are between them running MSLIC