Created SLIC and MSLIC
Using the previous changes writting a MSLIC wrapper was relativly strait forward. Changed the Name of kmeans_local to SLIC after altering the function structure a little as this is a more locical name and because the function change breaks the tests. \n Need to write new tests for the new two files and need to experement with different combination metrics and distance metrics.