First attempt at merge wrapper class
relaised a less complex method with more redunded calculations should be used here as the complexity has too many bug potentials and is too hard to code. Since the total number of segments here should never be large compared to the kmeans calculation run time will hopefully not be an issue here. If it is I can work on optimisations later ('premature optimisation is the root of all evil')